test(@ngtools/webpack): add unit test for register_locate_data transformer
