+++ packages/@ngtools/webpack/src/transformers/register_locale_data.ts
@@ -1,5 +1,3 @@
-import * as path from 'path';
-import * as fs from 'fs';
 import * as ts from 'typescript';
 
 import { findAstNodes, getFirstNode } from './ast_helpers';
@@ -45,39 +43,6 @@ export function registerLocaleData(
       return;
     }
 
-    // get the path of the common module
-    const commonPath = path.dirname(require.resolve('@angular/common/package.json'));
-    // check if the locale file exists
-    if (!fs.existsSync(path.resolve(commonPath, 'locales', `${locale}.js`))) {
-      // check for an alternative locale (if the locale id was badly formatted)
-      const locales = fs.readdirSync(path.resolve(commonPath, 'locales'))
-        .filter(file => file.endsWith('.js'))
-        .map(file => file.replace('.js', ''));
-
-      let newLocale;
-      const normalizedLocale = locale.toLowerCase().replace(/_/g, '-');
-      for (const l of locales) {
-        if (l.toLowerCase() === normalizedLocale) {
-          newLocale = l;
-          break;
-        }
-      }
-
-      if (newLocale) {
-        locale = newLocale;
-      } else {
-        // check for a parent locale
-        const parentLocale = normalizedLocale.split('-')[0];
-        if (locales.indexOf(parentLocale) !== -1) {
-          locale = parentLocale;
-        } else {
-          throw new Error(
-            `Unable to load the locale data file "@angular/common/locales/${locale}", ` +
-            `please check that "${locale}" is a valid locale id.`);
-        }
-      }
-    }
-
     // Create the import node for the locale.
     const localeIdentifier = ts.createIdentifier(`__locale_${locale.replace(/-/g, '')}__`);
     const localeImportClause = ts.createImportClause(localeIdentifier, undefined);
@@ -92,7 +57,7 @@ export function registerLocaleData(
 
     // Create the import node for the registerLocaleData function.
     const regIdentifier = ts.createIdentifier(`registerLocaleData`);
-    const regImportSpecifier = ts.createImportSpecifier(regIdentifier, regIdentifier);
+    const regImportSpecifier = ts.createImportSpecifier(undefined, regIdentifier);
     const regNamedImport = ts.createNamedImports([regImportSpecifier]);
     const regImportClause = ts.createImportClause(undefined, regNamedImport);
     const regNewImport = ts.createImportDeclaration(undefined, undefined, regImportClause,
